滚动条滚动由参数float改为int

+++ EzUI/include/Control.h
@@ -114,7 +114,7 @@ namespace EzUI {
 		void RemoveEventNotify(int eventType);
 		void SetCursor(const EString& fileName);
 		virtual ScrollBar* GetScrollBar();
-		virtual int MoveScroll(float offset);
+		virtual int MoveScroll(int offset);
 		HCURSOR GetCursor();
 	public:
 		Control();
