+++ EzUI/include/VList.h
@@ -20,7 +20,7 @@ namespace EzUI {
 		virtual void ChildPainting(Controls& controls, PaintEventArgs& args)override;
 		virtual void RefreshScroll(const int& _maxBottom);
 		virtual ScrollBar* GetScrollBar() override;
-		virtual int MoveScroll(float offset) override;
+		virtual int MoveScroll(int offset) override;
 
 	};
 };
